requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
backoff==2.2.1
banks==2.2.0
beautifulsoup4==4.13.4
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==45.0.6
dataclasses-json==0.6.7
datasets==4.0.0
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
emoji==2.14.1
filelock==3.19.1
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.3.0
greenlet==3.2.4
griffe==1.12.1
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-text-splitters==0.3.9
langdetect==1.0.9
langsmith==0.4.14
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.13.2
llama-index-cli==0.5.0
llama-index-core==0.13.2
llama-index-embeddings-openai==0.5.0
llama-index-indices-managed-llama-cloud==0.9.2
llama-index-instrumentation==0.4.0
llama-index-llms-openai==0.5.4
llama-index-readers-file==0.5.2
llama-index-readers-llama-parse==0.5.0
llama-index-workflows==1.3.0
llama-parse==0.6.54
lxml==6.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
numpy==2.3.2
olefile==0.47
openai==1.100.2
orjson==3.11.2
packaging==25.0
pandas==2.2.3
pillow==11.3.0
pinecone-client==6.0.0
pinecone-plugin-interface==0.0.7
platformdirs==4.3.8
propcache==0.3.2
psutil==7.0.0
pyarrow==21.0.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyMuPDF==1.26.3
pypdf==6.0.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-iso639==2025.2.18
python-magic==0.4.27
python-oxmsg==0.0.2
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2025.7.34
requests==2.32.5
requests-toolbelt==1.0.0
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
sentence-transformers==5.1.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.43
striprtf==0.0.26
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.21.4
torch==2.8.0
tqdm==4.67.1
transformers==4.55.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
unstructured==0.18.13
unstructured-client==0.42.3
urllib3==2.5.0
webencodings==0.5.1
wrapt==1.17.3
xxhash==3.5.0
yarl==1.20.1
zstandard==0.24.0